+++ lib/src/bones_api_base.dart
@@ -2,17 +2,17 @@ import 'dart:async';
 import 'dart:convert' as dart_convert;
 import 'dart:typed_data';
 
-import 'package:bones_api/bones_api.dart';
 import 'package:collection/collection.dart';
 import 'package:reflection_factory/reflection_factory.dart';
 
 import 'bones_api_config.dart';
 import 'bones_api_extension.dart';
+import 'bones_api_utils.dart';
 
 /// Root class of an API.
 abstract class APIRoot {
   // ignore: constant_identifier_names
-  static const String VERSION = '1.0.21';
+  static const String VERSION = '1.0.22';
 
   static final Map<String, APIRoot> _instances = <String, APIRoot>{};